Updated CentOS init script

+++ scripts/centos/default.sh
@@ -27,44 +27,49 @@ chkconfig ntpd on
 sed -e "s/\e\[5~/\e\[A/g" /etc/inputrc > /tmp/inputrc
 sed -e "s/\e\[6~/\e[B/g" /tmp/inputrc > /etc/inputrc
 
-### Begin: Not install security for OpenVZ
-ip link | grep venet
-if [ $? -ne 0 ];
-then
-
-### Install BFD and APF
-mkdir secdistr && cd secdistr
-
-mkdir -p apf && cd apf
-wget http://www.rfxn.com/downloads/apf-current.tar.gz
-tar --strip-components=1 -xzf apf-current.tar.gz
-./install.sh
-cd ..
-
-cp /etc/apf/conf.apf /etc/apf/conf.apf.bkp
-perl -pi -e 's/DEVEL_MODE="1"/DEVEL_MODE="0"/g' /etc/apf/conf.apf
-
 if [ ! -e /etc/cron.d ]; then
     yum -y install cronie
 fi
 
-# For OpenVZ
+### Begin: Do not install security for OpenVZ
 ip link | grep venet
-if [ $? -eq 0 ];
-then
-    perl -pi -e 's/IFACE_IN="eth0"/IFACE_IN="venet0"/g' /etc/apf/conf.apf
-    perl -pi -e 's/IFACE_OUT="eth0"/IFACE_OUT="venet0"/g' /etc/apf/conf.apf
-    perl -pi -e 's/SET_MONOKERN="0"/SET_MONOKERN="1"/g' /etc/apf/conf.apf
+if [ $? -ne 0 ]; then
+    ### Install BFD and APF
+    mkdir secdistr && cd secdistr
+
+    if [ ! -e /etc/apf ]; then
+        echo "Install APF"
+
+        mkdir -p apf && cd apf
+        wget http://www.rfxn.com/downloads/apf-current.tar.gz
+        tar --strip-components=1 -xzf apf-current.tar.gz
+        ./install.sh
+
+        cp /etc/apf/conf.apf /etc/apf/conf.apf.bkp
+        perl -pi -e 's/DEVEL_MODE="1"/DEVEL_MODE="0"/g' /etc/apf/conf.apf
+
+        # For OpenVZ
+        ip link | grep venet
+        if [ $? -eq 0 ];
+        then
+            perl -pi -e 's/IFACE_IN="eth0"/IFACE_IN="venet0"/g' /etc/apf/conf.apf
+            perl -pi -e 's/IFACE_OUT="eth0"/IFACE_OUT="venet0"/g' /etc/apf/conf.apf
+            perl -pi -e 's/SET_MONOKERN="0"/SET_MONOKERN="1"/g' /etc/apf/conf.apf
+        fi
+
+        echo "10.0.0.0/24" >> /etc/apf/allow_hosts.rules
+
+        cd ..
+    fi
+
+    if [ ! -e /usr/local/bfd ]; then
+        echo "Install BFD"
+
+        mkdir -p bfd && cd bfd
+        wget http://www.rfxn.com/downloads/bfd-current.tar.gz
+        tar --strip-components=1 -xzf bfd-current.tar.gz
+        ./install.sh
+        cd ..
+    fi
 fi
-
-echo "10.0.0.0/24" >> /etc/apf/allow_hosts.rules
-
-
-mkdir -p bfd && cd bfd
-wget http://www.rfxn.com/downloads/bfd-current.tar.gz
-tar --strip-components=1 -xzf bfd-current.tar.gz
-./install.sh
-cd ..
-
-fi
-### End: Not install security for OpenVZ
+### End: Do not install security for OpenVZ
